refactor: switch from package by layer to package by feature.
