+++ src/main/java/com/csdm/adbooker/MyAdBookerApplication.java
@@ -1,10 +1,7 @@
 package com.csdm.adbooker;
 
-import com.csdm.adbooker.service.NewsItemService;
-import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
 
 @SpringBootApplication
 public class MyAdBookerApplication {
