+++ src/main/java/com/csdm/adbooker/newsItem/NewsItemRepository.java
@@ -1,10 +1,8 @@
-package com.csdm.adbooker.repository;
+package com.csdm.adbooker.newsItem;
 
-import com.csdm.adbooker.model.NewsItem;
+import com.csdm.adbooker.newsItem.model.NewsItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
-import java.util.List;
 
 @Repository
 public interface NewsItemRepository extends JpaRepository<NewsItem, Long> {
